src/entities/Counter/ui/Counter.test.tsx
import { componentRender } from 'shared/lib/tests/componentsRender/componentRender';
import { screen } from '@testing-library/react';
import { userEvent } from '@storybook/testing-library';
import { Counter } from './Counter';

describe('Counter', () => {
    test('has on page', () => {
        componentRender(<Counter />, {
            initialState: { counter: { value: 10 } },
        });
        expect(screen.getByTestId('value-title-counter')).toHaveTextContent('10');
    });

    test('increment btn', () => {
        componentRender(<Counter />, {
            initialState: { counter: { value: 10 } },
        });
        userEvent.click(screen.getByTestId('increment-btn'));
        expect(screen.getByTestId('value-title-counter')).toHaveTextContent('11');
    });

    test('decrement btn', () => {
        componentRender(<Counter />, {
            initialState: { counter: { value: 10 } },
        });
        userEvent.click(screen.getByTestId('decrement-btn'));
        expect(screen.getByTestId('value-title-counter')).toHaveTextContent('9');
    });
});
